sys: util_macro: pairwise list iteration
Add a new family of argument list iteration macros, based on
`FOR_EACH_*`, that instead of calling a macro on every argument, calls
a macro on each pair of arguments. Naming follows the same convention
as `FOR_EACH_*`, prefixed with `PAIRWISE_`. Resulting in
`PAIRWISE_FOR_EACH(...)`, etc.

The primary value in this family of macros is the ability to iterate
over values of multiple types, in particular parameterizing a macro
argument:
```
 #define M1(arg) {.const = "x", .val = arg}
 #define M2(arg) {.const = "y", .val = arg}
 #define M3(arg) {.const = "z", .val = arg}

 #define M_EVAL(macro, arg) macro(arg)

struct t some_struct[] = {
    PAIRWISE_FOR_EACH(M_EVAL, (,) M1, 1, M2, 12, M3, -10),
};
```

The above functionality is not possible with the existing `FOR_EACH_*`
family of macros without defining an individual wrapper for each of
`M1`, `M2` and `M3` that contains the variable value.

Signed-off-by: Jordan Yates <jordan@embeint.com>