bluetooth: conn: add `conn` to recycled callback
Add the connection object that was recycled to the recycled callback.
This can be used to match the connections created from advertising sets
to the recycled callback invocation.

In turn, this allows software libraries to manage only a single
advertising set, restarting the set when a connection created from that
set terminates, emulating the behaviour of `!BT_LE_ADV_OPT_ONE_TIME`
for extended advertising.

Signed-off-by: Jordan Yates <jordan@embeint.com>