zbus: remove `k_malloc` dependency for `ZBUS_RUNTIME_OBSERVERS`
Remove the dependency on the system heap existing when enabling
`ZBUS_RUNTIME_OBSERVERS`. This means the overhead of the runtime list
is applied to all observers in the system, instead of just the observers
that are added at runtime. However this overhead is only 8 bytes per
observer (2 pointers). It is unlikely that this overhead is larger
than the cost of enabling the system heap for applications that don't
already have it enabled.

Signed-off-by: Jordan Yates <jordan@embeint.com>