spi: add `spi_acquire` call
Add a counterpart to the `spi_release` function call that locks the
context without needing to initiate a transaction. This is intended to
be used when SPI bus access needs to be blocked before the device is
accessed.

One example of this requirement is power cycling an SD card, where the
bus should be kept idle after power up until starting the switch from
SD mode to SPI mode. This requires locking the bus before power is
applied.

Signed-off-by: Jordan Yates <jordan@embeint.com>