sdhc: spi: power up with bus locked
Lock the SPI bus from access by other devices before applying power.
Keep the bus locked until the initial 74 clocks and CMD0 are sent to the
card. This prevents the card from incorrectly deciding that it should be
in SD mode due to bus traffic to other devices while it is going through
the init sequence.

Signed-off-by: Jordan Yates <jordan@embeint.com>