modem: modem_cellular: one-time ready script
Add a one-time optional script that is run when the modem moves into the
waiting for registration state, i.e. when the periodic script has been
scheduled but not yet run. This level can be used for commands that
should be run once the CMUX DLCI channels have been finalised.

Signed-off-by: Jordan Yates <jordan@embeint.com>