libc: picolibc: full path to toolchain `picolibc.specs`
For compilers that are known to expose the `-print-file-name` argument,
query the compiler for the full path to the `picolibc.spec` file that
is used for the `--specs=` argument.

The use of the relative `picolibc.spec` filename in the compiler
arguments causes `ccache` to immediately give up on attempting to cache
the compiler call, with internal debug errors like this:
```
Trying direct lookup
Failed to lstat picolibc.specs: No such file or directory
While processing -specs=picolibc.specs: picolibc.specs is missing
Failed; falling back to running the real compiler
```

In practice, this makes ccache unavailable for any application/test that
uses the Zephyr SDK picolibc module.

A more general solution than implemented in `d3f9f391`, which only fixes
the twister CI action (and codecov/clang actions in other commits).

Signed-off-by: Jordan Yates <jordan@embeint.com>