twister: runner: sequential run option
Add the option for tests to specify they must be run sequentially, i.e.
not in parallel with any other tests. This can be used for tests that
use some common resource on the host that cannot be shared between
multiple users.

The triggering use case are tests that use the native socket abstraction
to attempt to `bind` to a specific address on the host. This works fine
as long as only one test is attempting to bind at a time, but fails
with multiple users.

This is implemented by restricting the run operation to only execute on
the first worker spun up by the runner library for testcases specifying
`run_sequential: true`.

Signed-off-by: Jordan Yates <jordan@embeint.com>