kernel: indirect executable RAM symbol
Instead of marking RAM regions as executable when `CONFIG_XIP=n`, add
a dedicated symbol to control the behaviour. XIP is the most obvious
case that requires this, but LLEXT for example can also load code into
RAM that is then executed.

The addition of `XIP_NOT` forces `EXECUTABLE_RAM` when `XIP` is disabled
in order to prevent users from accidently turning it off.

Signed-off-by: Jordan Yates <jordan@embeint.com>