lora: preamble and header detection callbacks
Add the option to receive callbacks when the LoRa modem detects a valid
preamble or packet header. This can be used to detect whether the modem
is in the middle of receiving a packet, allowing the user to possibly
defer a mode switch (to TX or IDLE) if it is. This is useful due to the
generally long packet durations on LoRa.

Signed-off-by: Jordan Yates <jordan@embeint.com>